Make keynote bench job reusable
