+++ Cargo.toml
@@ -66,6 +66,7 @@ members = [
   "sdks/rust/tests/procedural-view-pk-client",
   "sdks/rust/tests/event-table-client",
   "tools/ci",
+  "tools/keynote-bench-harness",
   "tools/license-check",
   "tools/replace-spacetimedb",
   "tools/generate-client-api",
