+++ tools/ci/Cargo.toml
@@ -15,4 +15,5 @@ serde_json.workspace = true
 duct.workspace = true
 tempfile.workspace = true
 env_logger.workspace = true
+keynote-bench-harness = { path = "../keynote-bench-harness" }
 spacetimedb-guard.workspace = true
