+++ tools/keynote-bench-harness/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "keynote-bench-harness"
+version = "0.1.0"
+edition.workspace = true
+publish = false
+
+[dependencies]
+anyhow.workspace = true
+reqwest = { workspace = true, features = ["blocking"] }
+serde_json.workspace = true
+tempfile.workspace = true
